requirements.txt
line-bot-sdk
google-cloud-bigquery